requirements-dev.txt
--requirement=requirements.txt

ipdb>=0.7
django-debug-toolbar>=0.9.4
pylint>=0.25
factory_boy==1.2.0
